terraform/consumer/requirements.txt
kafka-python==2.0.2
slack-sdk==3.21.3
pymsteams==0.2.2
python-dotenv==1.0.0
requests==2.31.0
boto3==1.28.60
aws-msk-iam-sasl-signer-python==1.0.2